+++ components/script/dom/cssviewportrule.rs
@@ -12,7 +12,7 @@ use dom::window::Window;
 use dom_struct::dom_struct;
 use style::shared_lock::{Locked, ToCssWithGuard};
 use style::stylearc::Arc;
-use style::viewport::ViewportRule;
+use style::stylesheets::ViewportRule;
 
 #[dom_struct]
 pub struct CSSViewportRule {
