+++ components/style/lib.rs
@@ -101,7 +101,6 @@ pub mod context;
 pub mod counter_style;
 pub mod custom_properties;
 pub mod data;
-pub mod document_condition;
 pub mod dom;
 pub mod element_state;
 #[cfg(feature = "servo")] mod encoding_support;
@@ -111,7 +110,6 @@ pub mod font_metrics;
 #[cfg(feature = "gecko")] #[allow(unsafe_code)] pub mod gecko;
 #[cfg(feature = "gecko")] #[allow(unsafe_code)] pub mod gecko_bindings;
 pub mod invalidation;
-pub mod keyframes;
 #[allow(missing_docs)] // TODO.
 pub mod logical_geometry;
 pub mod matching;
@@ -134,14 +132,12 @@ pub mod style_adjuster;
 pub mod stylearc;
 pub mod stylesheet_set;
 pub mod stylesheets;
-pub mod supports;
 pub mod thread_state;
 pub mod timer;
 pub mod traversal;
 #[macro_use]
 #[allow(non_camel_case_types)]
 pub mod values;
-pub mod viewport;
 
 use std::fmt;
 use style_traits::ToCss;
